+++ changes/77.feature.md
@@ -1,7 +1,1 @@
-
-Features
-
-- Added schedule plotting, with theme options similar to that of the circuit. 
-- Adding schedule style classes for added customizability 
-- Adding ``LinearSchedule`` which implements linear interpolation between the user defined coefficient during the schedule.
-- Updated ``Schedule`` and ``LinearSchedule`` to enforce that T is divisible by dt.
+Implemented schedule plotting with theme options consistent with those available for circuits, along with new schedule style classes to improve customization. A new `LinearSchedule` has been added, implementing linear interpolation between user-defined coefficients across the schedule. In addition, both `Schedule` and `LinearSchedule` now enforce that the total duration `T` must be divisible by the time step `dt`.
