+++ src/qilisdk/analog/schedule.py
@@ -53,12 +53,12 @@ class Schedule(Parameterizable):
         Raises:
             ValueError: If the provided schedule references Hamiltonians that have not been defined.
         """
-        if abs(T % dt) > 1e-12:  # noqa: PLR2004
-            raise ValueError("T must be divisible by dt.")
         if not isinstance(T, int):
             raise ValueError("T must be an integer")
         if not isinstance(dt, int):
             raise ValueError("dt must be an integer")
+        if abs(T % dt) != 0:
+            raise ValueError("T must be divisible by dt.")
         self._hamiltonians: dict[str, Hamiltonian] = hamiltonians if hamiltonians is not None else {}
         self._schedule: dict[int, dict[str, float | Term]] = schedule if schedule is not None else {0: {}}
         self._parameters: dict[str, Parameter] = {}
@@ -500,7 +500,7 @@ class Schedule(Parameterizable):
             return result
         raise StopIteration
 
-    def draw(self, style: ScheduleStyle | None = None, filepath: str | None = None, title: str | None = None) -> None:
+    def draw(self, style: ScheduleStyle | None = None, filepath: str | None = None) -> None:
         """Render a plot of the schedule using matplotlib and optionally save it to a file.
 
         The schedule is rendered using the provided style configuration. If ``filepath`` is
@@ -516,106 +516,6 @@ class Schedule(Parameterizable):
 
         style = style or ScheduleStyle()
         renderer = MatplotlibScheduleRenderer(self, style=style)
-        renderer.plot(title=title)
+        renderer.plot()
         if filepath:
             renderer.save(filepath)
-
-
-@yaml.register_class
-class LinearSchedule(Schedule):
-    """A Schedule that linearly interpolates between defined time steps."""
-
-    def get_coefficient(self, time_step: float, hamiltonian_key: str) -> Number:
-        t = time_step / self.dt
-        t_idx = int(t)
-
-        # if exactly defined, return directly
-        if t_idx in self._schedule and hamiltonian_key in self._schedule[t_idx]:
-            val = self._schedule[t_idx][hamiltonian_key]
-            return (
-                val.evaluate({}) if isinstance(val, Term) else (val.evaluate() if isinstance(val, Parameter) else val)
-            )
-
-        # search backwards for last defined
-        prev_idx, prev_val = None, None
-        for i in range(t_idx, -1, -1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                prev_idx = i
-                val = self._schedule[i][hamiltonian_key]
-                prev_val = (
-                    val.evaluate({})
-                    if isinstance(val, Term)
-                    else (val.evaluate() if isinstance(val, Parameter) else val)
-                )
-                break
-
-        # search forwards for next defined
-        next_idx, next_val = None, None
-        for i in range(t_idx + 1, int(self.T / self.dt) + 1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                next_idx = i
-                val = self._schedule[i][hamiltonian_key]
-                next_val = (
-                    val.evaluate({})
-                    if isinstance(val, Term)
-                    else (val.evaluate() if isinstance(val, Parameter) else val)
-                )
-                break
-
-        # cases
-        if prev_val is None and next_val is None:
-            return 0
-        if prev_val is None and next_val is not None:
-            return next_val
-        if next_val is None and prev_val is not None:
-            return prev_val
-
-        # linear interpolation
-        if next_idx is None or prev_idx is None or prev_val is None or next_val is None:
-            raise ValueError("Something unexpected happened while retrieving the coefficient.")
-        alpha: float = (t - prev_idx) / (next_idx - prev_idx)
-        return (1 - alpha) * prev_val + alpha * next_val
-
-    def get_coefficient_expression(self, time_step: float, hamiltonian_key: str) -> Number | Term:
-        t = time_step / self.dt
-        t_idx = int(t)
-
-        if t_idx in self._schedule and hamiltonian_key in self._schedule[t_idx]:
-            return self._schedule[t_idx][hamiltonian_key]
-
-        # search backwards
-        prev_idx, prev_expr = None, None
-        for i in range(t_idx, -1, -1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                prev_idx = i
-                prev_expr = self._schedule[i][hamiltonian_key]
-                break
-
-        # search forwards
-        next_idx, next_expr = None, None
-        for i in range(t_idx + 1, int(self.T / self.dt) + 1):
-            if i in self._schedule and hamiltonian_key in self._schedule[i]:
-                next_idx = i
-                next_expr = self._schedule[i][hamiltonian_key]
-                break
-
-        # cases
-        if prev_expr is None and next_expr is None:
-            return 0
-        if prev_expr is None and next_expr is not None:
-            return next_expr
-        if next_expr is None and prev_expr is not None:
-            return prev_expr
-
-        # linear interpolation (keeps expressions if they are Terms/Parameters)
-        if next_idx is None or prev_idx is None or prev_expr is None or next_expr is None:
-            raise ValueError("Something unexpected happened while retrieving the coefficient.")
-        alpha: float = (t - prev_idx) / (next_idx - prev_idx)
-        return (1 - alpha) * prev_expr + alpha * next_expr
-
-    def __getitem__(self, time_step: int) -> Hamiltonian:
-        ham = Hamiltonian()
-        for ham_label in self._hamiltonians:
-            coeff = self.get_coefficient(time_step * self.dt, ham_label)
-            ham += coeff * self._hamiltonians[ham_label]
-        return ham.get_static_hamiltonian()
